新增Roam表，预加入漫游功能。 支持杜比和8K视频（理论上之前也可以） 引入BRv 修改解析策略，点击item进入解析，而非采用点击button 修改SD储存目录，不允许随意改动

+++ app/src/main/java/com/imcys/bilibilias/home/ui/model/VideoPlayBean.java
@@ -243,7 +243,7 @@ public class VideoPlayBean  implements Serializable {
 
             private int order;
             private int length;
-            private int size;
+            private Long size;
             private String ahead;
             private String vhead;
             private String url;
@@ -265,11 +265,11 @@ public class VideoPlayBean  implements Serializable {
                 this.length = length;
             }
 
-            public int getSize() {
+            public Long getSize() {
                 return size;
             }
 
-            public void setSize(int size) {
+            public void setSize(Long size) {
                 this.size = size;
             }
